requirements.txt
fastapi>=0.70.0
pydantic>=1.8.2
python-dotenv>=1.0.0
uvicorn[standard]>=0.15.0
langchain-openai>=0.0.8
langchain-community>=0.0.20
langchain-core>=0.1.33
psycopg2-binary>=2.9.9
openai>=1.14.0
sqlalchemy>=2.0.25